MAJ suite MAJ GALGAS

+++ hand_coded_sources/C_us.h
@@ -80,7 +80,7 @@ class C_us {
   public : static void reallocUniqueTable (const int32_t inTableUniqueNewSize) ;
 
 //--- Realloc addition cache
-  public : static void reallocAdditionCache (const int32_t inNewCacheSize) ;
+  public : static void reallocAdditionCache (const size_t inNewCacheSize) ;
 
 //--- Decision list
   protected : C_us_nodeInfo * mRootPointer ;
